requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
sqlmodel
alembic
pydantic-settings
python-multipart
passlib[bcrypt]
python-jose[cryptography]
pytest
httpx
email-validator
aiofiles
prometheus-fastapi-instrumentator
pytest-asyncio
pytest-cov
redis
